pyproject.toml
[tool.poetry]
name = "fastapi-di-auth-template"
version = "0.1.0"
description = ""
authors = ["Yuchan Han <serltretu24@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.8"}
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
uvicorn = "^0.34.0"
fastapi-restful = {extras = ["all"], version = "^0.6.0"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.24.0"}
dependency-injector = "^4.45.0"
aiohttp = "^3.11.12"
aiogoogle = "^5.13.2"
pyjwt = "^2.10.1"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
